Move PartyField to parties.fields
